requirements.txt
celery==5.3.1
Django==4.2.4
django-cors-headers==4.1.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
google-auth==2.21.0
Pillow==10.0.0
python-dotenv==1.0.0
redis==4.6.0
stripe==5.4.0
drf-yasg==1.21.7
jwt==1.3.1
moviepy==1.0.3
